+++ CHANGES.md
@@ -11,6 +11,11 @@
 
 ## develop
 
+- [ADD] 音声ルート変更イベントとして `SoraHandlers.onChangeAudioRoute` を追加する
+  - `RTCAudioSessionDelegate.audioSessionDidChangeRoute` を利用してコールバックする
+  - コールバック引数として `RTCAudioSession`, `reason`, `previousRoute` を渡す
+  - 音声ルートの状態などの評価はコールバック内で `RTCAudioSession.currentRoute` を参照する
+  - @t-miya
 - [ADD] iOS 端末画面をキャプチャして配信する ScreenCapture を追加する
   - MediaChannel に画面キャプチャ開始 / 停止 API を追加する
   - 画面キャプチャ開始時に渡す設定として `ScreenCaptureSettings` 構造体を追加する